backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Resume processing
PyMuPDF==1.23.8
python-docx==0.8.11

# Face recognition
face-recognition==1.3.0
numpy==1.24.3

# Production server
gunicorn==21.2.0
